Move project runner implementation into internal/projectrunner
Promote the 12-line shim to the real runner package: the 17 runner-only
files and their tests move out of internal/cli, and Run now calls
RunProjectLocal in the same package. Cross-package test helpers are
duplicated with a note instead of widening the production API.
go list -deps confirms internal/cli and internal/projectrunner no
longer reference each other.
