+++ cli/internal/projectrunner/compile_wait.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"context"
