+++ cli/internal/projectrunner/connection_retry.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"context"
