+++ cli/internal/projectrunner/connection_retry_test.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"bufio"
