+++ cli/internal/projectrunner/control_play_mode_wait.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"context"
