+++ cli/internal/projectrunner/control_play_mode_wait_test.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"bufio"
