+++ cli/internal/projectrunner/debug_timing.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"encoding/json"
