+++ cli/internal/projectrunner/debug_timing_test.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"bytes"
