+++ cli/internal/projectrunner/dynamic_code_file.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"fmt"
