+++ cli/internal/projectrunner/dynamic_code_wait.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"encoding/json"
