+++ cli/internal/projectrunner/execution_errors.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"fmt"
