+++ cli/internal/projectrunner/list_output.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"encoding/json"
