+++ cli/internal/projectrunner/list_output_test.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"encoding/json"
