+++ cli/internal/projectrunner/pause_point_errors.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"fmt"
