+++ cli/internal/projectrunner/pause_point_wait.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"context"
