+++ cli/internal/projectrunner/runner_commands.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"context"
