+++ cli/internal/projectrunner/runner_usage.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"io"
