+++ cli/internal/projectrunner/tool_params.go
@@ -1,4 +1,4 @@
-package cli
+package projectrunner
 
 import (
 	"encoding/json"
